+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fix for compatibility with Alpine Linux.
+- Fix deserialization of `rel` arrays.
 
 ## [1.30.0] - 2023-07-18